chore(@clayui/pagination-bar): update PaginationBar component stories
